log de proyectos y usuarios v.2

+++ src/proyecto/Proyecto.java
@@ -53,8 +53,9 @@ public final class Proyecto {
         return cantidadNecesaria;
     }
 
-    public double getCantidadFinanciada() {
-        return cantidadFinanciada;
+    public int getCantidadFinanciada() {
+
+        return (int)cantidadFinanciada;
     }
 
     public LocalDate getFechaInicio() {
